add: types for special plugin_specs object
config.mapping.KeyMap.PluginSpecs is an object that will be used to
properly set up keymaps related to a special set of plugins.

These plugins are special because they demand a special format to
register keymaps in their configuration. In other words: they are
plugins in which their associated keymaps are not registered with the
conventional vim.keymap.set(...).

At the time of this commit, the only plugins that require this special
setup are nvim-treesitter and nvim-treesitter-textobjects