refactor: change xf86 brightness keybinds to $mod+f5|f6
I resorted to this because Fn+F5 and Fn+F6 are not recognized by my
current operating system (Ubuntu 24.04) for some odd reason.

Therefore, my only option is to use $mod (Windows key) instead of Fn. If
I happen to resolve this issue, I should be able to revert back to the
XF86 keybinds.